pyproject.toml
# 文件: pyproject.toml（仅示例，合并到你的现有配置中）
[project]
name = "langextract_demo"
version = "0.1.0"
description = "LangExtract 医疗场景抽取示例"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langextract[openai]>=1.0.8",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
]

[tool.uv]
# 可选：锁定源或其他 uv 配置
